pdfviewer/requirements.txt
streamlit>=1.32.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
Pillow>=10.0.0
